requirements.txt
altair==5.5.0
attrs==25.4.0
audioread==3.1.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.10.2
networkx==3.5
nltk==3.9.2
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==6.33.0
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
